Delete contact list from overview page
