+++ locales/de-DE/translation.json
@@ -304,6 +304,7 @@
   "See more": "See more",
   "Latest transactions": "Latest transactions",
   "Addresses": "Adressen",
+  "Your addresses": "Ihre Adressen",
   "Developer tools": "Developer tools",
   "Enable developer tools": "Enable developer tools",
   "Features for developers only": "Features for developers only",
