+++ locales/en-US/translation.json
@@ -305,6 +305,7 @@
   "See more": "See more",
   "Latest transactions": "Latest transactions",
   "Addresses": "Addresses",
+  "Your addresses": "Your addresses",
   "Developer tools": "Developer tools",
   "Enable developer tools": "Enable developer tools",
   "Features for developers only": "Features for developers only",
